Refactor : separate module for playing game
